+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
-## [Unreleased] - 2024-05-03
+## [1.3.0] - 2024-05-03
 
 ### Added
 
@@ -326,3 +326,4 @@ and this project adheres to [Semantic Versioning].
 [1.1.8]: https://github.com/LumiRadio/lumiRadio/compare/v1.1.7..v1.1.8
 [1.1.9]: https://github.com/LumiRadio/lumiRadio/compare/v1.1.8..v1.1.9
 [1.2.0]: https://github.com/LumiRadio/lumiRadio/compare/v1.1.9..v1.2.0
+[1.3.0]: https://github.com/LumiRadio/lumiRadio/compare/v1.2.0..v1.3.0